requirements.mcp.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
email-validator==2.1.1
fastapi==0.135.3
graphql-core==3.2.8
greenlet==3.4.0
h11==0.16.0
hiredis==3.3.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
limits==5.8.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.27.0
numpy==2.4.4
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
PyWavelets==1.9.0
prometheus_client==0.21.0
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
strawberry-graphql==0.243.0
stripe==15.0.1
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
ariadne==1.0.1
aiohttp==3.11.18
posthog==3.7.0